+++ .prettierignore
@@ -1,0 +1,7 @@
+node_modules/
+.idea/
+.vscode/
+build/
+dist/
+coverage/
+.env
