+++ app.js
@@ -1,33 +1,31 @@
-const express = require('express')
-const logger = require('morgan')
-const cors = require('cors')
+const express = require('express');
+const logger = require('morgan');
+const cors = require('cors');
 
-const { routerContacts } = require('./routes/api/contacts')
+const { routerContacts } = require('./routes/api/contacts');
 
-const app = express()
+const app = express();
 
-const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short'
+const formatsLogger = app.get('env') === 'development' ? 'dev' : 'short';
 
-app.use(logger(formatsLogger))
-app.use(cors())
-app.use(express.json())
-
+app.use(logger(formatsLogger));
+app.use(cors());
+app.use(express.json());
 
 // routes
 
-app.use('/api/contacts', routerContacts )
+app.use('/api/contacts', routerContacts);
 
 // 404
 
 app.use((req, res) => {
-  res.status(404).json({ message: 'Not found' })
-})
+  res.status(404).json({ message: 'Not found' });
+});
 
 // error handling
 
 app.use((err, req, res, next) => {
-  res.status(500).json({ message: err.message })
-})
+  res.status(500).json({ message: err.message });
+});
 
-
-module.exports = app
+module.exports = app;
