+++ helpers/index.js
@@ -1,0 +1,20 @@
+function tryCatchWrapper(endpointFn) {
+  return async (req, res, next) => {
+    try {
+      await endpointFn(req, res, next);
+    } catch (error) {
+      return next(error);
+    }
+  };
+}
+
+function HttpError(status, message) {
+  const err = new Error(message);
+  err.status = status;
+  return err;
+}
+
+module.exports = {
+  tryCatchWrapper,
+  HttpError,
+};
