--- models/helpers/index.js
@@ -1,25 +1,0 @@
- function tryCatchWrapper(enpointFn) {
-    return async (req, res, next) => {
-        try{
-            await enpointFn(req, res, next);
-        }catch (error) {
-            return next(error);
-        }
-    };
-}
-
-
-function HttpError(status, message ) {
-    const err = new Error(message)
-    err.status = status
-    return err;
-
-
-}
-
-
-
-module.exports = {
-    tryCatchWrapper,
-    HttpError,
-}
