+++ public/about.html
@@ -4,14 +4,15 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home Page</title>
+    <title>About</title>
 </head>
 <body>
-    My Movies Home Page
+About
 
-    <div>
-        <a href="./about.html">go to about</a>
-    </div>
-    
+<div>
+    <a href="/public">go to home</a>
+</div>
+
+
 </body>
 </html>
