+++ public/index.html
@@ -4,15 +4,14 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>About</title>
+    <title>Home Page</title>
 </head>
 <body>
-About
+    My Movies Home Page
 
-<div>
-    <a href="/">go to home</a>
-</div>
-
-
+    <div>
+        <a href="about.html">go to about</a>
+    </div>
+    
 </body>
 </html>
